+++ integrations/langfuse/src/haystack_integrations/components/connectors/langfuse/langfuse_connector.py
@@ -21,21 +21,18 @@ class LangfuseConnector:
     LangfuseConnector connects Haystack LLM framework with [Langfuse](https://langfuse.com) in order to enable the
     tracing of operations and data flow within various components of a pipeline.
 
-    Simply add this component to your pipeline, but *do not* connect it to any other component. The LangfuseConnector
-    will automatically trace the operations and data flow within the pipeline.
+    To use LangfuseConnector, add it to your pipeline without connecting it to any other components.
+    It will automatically trace all pipeline operations when tracing is enabled.
 
-    Note that you need to set the `LANGFUSE_SECRET_KEY` and `LANGFUSE_PUBLIC_KEY` environment variables in order
-    to use this component. The `LANGFUSE_SECRET_KEY` and `LANGFUSE_PUBLIC_KEY` are the secret and public keys provided
-    by Langfuse. You can get these keys by signing up for an account on the Langfuse website.
+    **Environment Configuration:**
+    - `LANGFUSE_SECRET_KEY` and `LANGFUSE_PUBLIC_KEY`: Required Langfuse API credentials.
+    - `HAYSTACK_CONTENT_TRACING_ENABLED`: Must be set to `"true"` to enable tracing.
+    - `HAYSTACK_LANGFUSE_ENFORCE_FLUSH`: (Optional) If set to `"false"`, disables flushing after each component.
+      Be cautious: this may cause data loss on crashes unless you manually flush before shutdown.
+      By default, the data is flushed after each component and blocks the thread until the data is sent to Langfuse.
 
-    In addition, you need to set the `HAYSTACK_CONTENT_TRACING_ENABLED` environment variable to `true` in order to
-    enable Haystack tracing in your pipeline.
-
-    Lastly, you may disable flushing the data after each component by setting the `HAYSTACK_LANGFUSE_ENFORCE_FLUSH`
-    environment variable to `false`. By default, the data is flushed after each component and blocks the thread until
-    the data is sent to Langfuse. **Caution**: Disabling this feature may result in data loss if the program crashes
-    before the data is sent to Langfuse. Make sure you will call langfuse.flush() explicitly before the program exits.
-    E.g. by using tracer.actual_tracer.flush():
+    If you disable flushing after each component make sure you will call langfuse.flush() explicitly before the
+    program exits. For example:
 
     ```python
     from haystack.tracing import tracer
@@ -51,13 +48,12 @@ class LangfuseConnector:
 
     # ...
 
-
     @app.on_event("shutdown")
     async def shutdown_event():
         tracer.actual_tracer.flush()
     ```
 
-    Here is an example of how to use it:
+    Here is an example of how to use LangfuseConnector in a pipeline:
 
     ```python
     import os
@@ -72,37 +68,36 @@ class LangfuseConnector:
         LangfuseConnector,
     )
 
-    if __name__ == "__main__":
-        pipe = Pipeline()
-        pipe.add_component("tracer", LangfuseConnector("Chat example"))
-        pipe.add_component("prompt_builder", ChatPromptBuilder())
-        pipe.add_component("llm", OpenAIChatGenerator(model="gpt-3.5-turbo"))
+    pipe = Pipeline()
+    pipe.add_component("tracer", LangfuseConnector("Chat example"))
+    pipe.add_component("prompt_builder", ChatPromptBuilder())
+    pipe.add_component("llm", OpenAIChatGenerator(model="gpt-4o-mini"))
 
-        pipe.connect("prompt_builder.prompt", "llm.messages")
+    pipe.connect("prompt_builder.prompt", "llm.messages")
 
-        messages = [
-            ChatMessage.from_system(
-                "Always respond in German even if some input data is in other languages."
-            ),
-            ChatMessage.from_user("Tell me about {{location}}"),
-        ]
+    messages = [
+        ChatMessage.from_system(
+            "Always respond in German even if some input data is in other languages."
+        ),
+        ChatMessage.from_user("Tell me about {{location}}"),
+    ]
 
-        response = pipe.run(
-            data={
-                "prompt_builder": {
-                    "template_variables": {"location": "Berlin"},
-                    "template": messages,
-                }
+    response = pipe.run(
+        data={
+            "prompt_builder": {
+                "template_variables": {"location": "Berlin"},
+                "template": messages,
             }
-        )
-        print(response["llm"]["replies"][0])
-        print(response["tracer"]["trace_url"])
-        print(response["tracer"]["trace_id"])
+        }
+    )
+    print(response["llm"]["replies"][0])
+    print(response["tracer"]["trace_url"])
+    print(response["tracer"]["trace_id"])
     ```
 
-    For advanced use cases, you can also customize how spans are created and processed by
-    providing a custom SpanHandler. This allows you to add custom metrics, set warning levels,
-    or attach additional metadata to your Langfuse traces:
+    For advanced use cases, you can also customize how spans are created and processed by providing a custom
+    SpanHandler. This allows you to add custom metrics, set warning levels, or attach additional metadata to your
+    Langfuse traces:
 
     ```python
     from haystack_integrations.tracing.langfuse import DefaultSpanHandler, LangfuseSpan
@@ -117,7 +112,6 @@ class LangfuseConnector:
 
     connector = LangfuseConnector(span_handler=CustomSpanHandler())
     ```
-
     """
 
     def __init__(
@@ -128,12 +122,15 @@ class LangfuseConnector:
         secret_key: Optional[Secret] = Secret.from_env_var("LANGFUSE_SECRET_KEY"),  # noqa: B008
         httpx_client: Optional[httpx.Client] = None,
         span_handler: Optional[SpanHandler] = None,
-    ):
+        *,
+        host: Optional[str] = None,
+        langfuse_client_kwargs: Optional[Dict[str, Any]] = None,
+    ) -> None:
         """
         Initialize the LangfuseConnector component.
 
-        :param name: The name of the pipeline or component. This name will be used to identify the tracing run on the
-            Langfuse dashboard.
+        :param name: The name for the trace. This name will be used to identify the tracing run in the Langfuse
+            dashboard.
         :param public: Whether the tracing data should be public or private. If set to `True`, the tracing data will be
             publicly accessible to anyone with the tracing URL. If set to `False`, the tracing data will be private and
             only accessible to the Langfuse account owner. The default is `False`.
@@ -146,18 +143,28 @@ class LangfuseConnector:
             The span handler controls how spans are created and processed, allowing customization of span types
             based on component types and additional processing after spans are yielded. See SpanHandler class for
             details on implementing custom handlers.
+        host: Host of Langfuse API. Can also be set via `LANGFUSE_HOST` environment variable.
+            By default it is set to `https://cloud.langfuse.com`.
+        :param langfuse_client_kwargs: Optional custom configuration for the Langfuse client. This is a dictionary
+            containing any additional configuration options for the Langfuse client. See the Langfuse documentation
+            for more details on available configuration options.
         """
         self.name = name
         self.public = public
         self.secret_key = secret_key
         self.public_key = public_key
         self.span_handler = span_handler
+        self.host = host
+        self.langfuse_client_kwargs = langfuse_client_kwargs
+        resolved_langfuse_client_kwargs = {
+            "secret_key": secret_key.resolve_value() if secret_key else None,
+            "public_key": public_key.resolve_value() if public_key else None,
+            "httpx_client": httpx_client,
+            "host": host,
+            **(langfuse_client_kwargs or {}),
+        }
         self.tracer = LangfuseTracer(
-            tracer=Langfuse(
-                secret_key=secret_key.resolve_value() if secret_key else None,
-                public_key=public_key.resolve_value() if public_key else None,
-                httpx_client=httpx_client,
-            ),
+            tracer=Langfuse(**resolved_langfuse_client_kwargs),
             name=name,
             public=public,
             span_handler=span_handler,
@@ -165,7 +172,7 @@ class LangfuseConnector:
         tracing.enable_tracing(self.tracer)
 
     @component.output_types(name=str, trace_url=str, trace_id=str)
-    def run(self, invocation_context: Optional[Dict[str, Any]] = None):
+    def run(self, invocation_context: Optional[Dict[str, Any]] = None) -> Dict[str, str]:
         """
         Runs the LangfuseConnector component.
 
@@ -191,6 +198,13 @@ class LangfuseConnector:
         :returns: The serialized component as a dictionary.
         """
         span_handler = serialize_class_instance(self.span_handler) if self.span_handler else None
+        if self.langfuse_client_kwargs:
+            # pop httpx_client and mask from self._langfuse_client_kwargs to prevent serialization issues
+            langfuse_client_kwargs = {
+                k: v for k, v in self.langfuse_client_kwargs.items() if k not in ["httpx_client", "mask"]
+            }
+        else:
+            langfuse_client_kwargs = None
         return default_to_dict(
             self,
             name=self.name,
@@ -199,6 +213,8 @@ class LangfuseConnector:
             public_key=self.public_key.to_dict() if self.public_key else None,
             # Note: httpx_client is not serialized as it's not serializable
             span_handler=span_handler,
+            host=self.host,
+            langfuse_client_kwargs=langfuse_client_kwargs,
         )
 
     @classmethod